Moved Checkers implementation code to CheckerAsserting.
